+++ plugins/backend/backstage-plugin-argo-cd-backend/src/service/argocd.test.ts
@@ -9,6 +9,7 @@ import fetchMock from 'jest-fetch-mock';
 import { timer } from './timer.services';
 import { mocked } from 'ts-jest/utils';
 import { Logger } from 'winston';
+import { UpdateArgoProjectAndAppProps } from './types';
 
 fetchMock.enableMocks();
 jest.mock('./timer.services');
@@ -123,6 +124,55 @@ describe('ArgoCD service', () => {
     ).rejects.toThrow();
   });
 
+  it('should get all argo app data if no option is provided', async () => {
+    fetchMock.mockResponseOnce(
+      JSON.stringify({
+        items: [
+          {
+            metadata: {
+              name: 'testAppName',
+              namespace: 'testNamespace',
+            },
+          },
+          {
+            metadata: {
+              name: 'testAppName2',
+              namespace: 'testNamespace2',
+            },
+          },
+        ],
+      }),
+    );
+    const resp = await argoService.getArgoAppData(
+      'https://argoInstance1.com',
+      'argoInstance1',
+      'testToken',
+    );
+
+    expect(resp).toStrictEqual({
+      items: [
+        {
+          metadata: {
+            name: 'testAppName',
+            namespace: 'testNamespace',
+            instance: {
+              name: 'argoInstance1',
+            },
+          },
+        },
+        {
+          metadata: {
+            name: 'testAppName2',
+            namespace: 'testNamespace2',
+            instance: {
+              name: 'argoInstance1',
+            },
+          },
+        },
+      ],
+    });
+  });
+
   it('should get argo app data', async () => {
     fetchMock.mockResponseOnce(
       JSON.stringify({
@@ -136,8 +186,8 @@ describe('ArgoCD service', () => {
     const resp = await argoService.getArgoAppData(
       'https://argoInstance1.com',
       'argoInstance1',
-      { name: 'testApp' },
       'testToken',
+      { name: 'testApp' },
     );
 
     expect(resp).toStrictEqual({
@@ -156,8 +206,8 @@ describe('ArgoCD service', () => {
       argoService.getArgoAppData(
         'https://argoInstance1.com',
         'argoInstance1',
-        { name: 'testApp' },
         'testToken',
+        { name: 'testApp' },
       ),
     ).rejects.toThrow();
   });
@@ -185,8 +235,8 @@ describe('ArgoCD service', () => {
     const resp = await argoService.getArgoAppData(
       'https://argoInstance1.com',
       'argoInstance1',
-      { selector: 'service=testApp' },
       'testToken',
+      { selector: 'service=testApp' },
     );
 
     expect(resp).toStrictEqual({
@@ -311,24 +361,23 @@ describe('ArgoCD service', () => {
   it('should fail to create an app in argo when argo errors out', async () => {
     fetchMock.mockResponseOnce(
       JSON.stringify({
-        error: 'Failed to Create application',
+        message: 'errorMessage',
       }),
+      { status: 500 },
     );
 
-    const resp = await argoService.createArgoApplication({
-      baseUrl: 'https://argoInstance1.com',
-      argoToken: 'testToken',
-      appName: 'testProject',
-      projectName: 'testProject',
-      namespace: 'test-namespace',
-      sourceRepo: 'https://github.com/backstage/backstage',
-      sourcePath: 'kubernetes/nonproduction',
-      labelValue: 'backstageId',
-    });
-
-    expect(resp).toStrictEqual({
-      error: 'Failed to Create application',
-    });
+    await expect(
+      argoService.createArgoApplication({
+        baseUrl: 'https://argoInstance1.com',
+        argoToken: 'testToken',
+        appName: 'testProject',
+        projectName: 'testProject',
+        namespace: 'test-namespace',
+        sourceRepo: 'https://github.com/backstage/backstage',
+        sourcePath: 'kubernetes/nonproduction',
+        labelValue: 'backstageId',
+      }),
+    ).rejects.toThrow('Error creating argo app: errorMessage');
   });
 
   it('should fail to create a application in argo when argo user is not given enough permissions', async () => {
@@ -999,6 +1048,149 @@ describe('ArgoCD service', () => {
 
       expect(loggerMock.error).toHaveBeenCalledWith(
         'Error getting token from Argo Instance argoInstance1: Error: FetchError',
+      );
+    });
+  });
+
+  describe('updateArgoProjectAndApp', () => {
+    const data: UpdateArgoProjectAndAppProps = {
+      instanceConfig: { name: 'argoInstanceName', url: 'url' },
+      argoToken: 'argoToken',
+      appName: 'appName',
+      projectName: 'projectName',
+      namespace: 'namespace',
+      sourceRepo: 'sourceRepo',
+      sourcePath: 'sourcePath',
+      labelValue: 'labelValue',
+    };
+    const getArgoAppDataResp = {
+      metadata: {
+        name: 'argoInstanceName',
+        resourceVersion: 'resourceVersion',
+      },
+      spec: { source: { repoURL: 'sourceRepo' } },
+      status: {},
+      instance: 'instance',
+    };
+    const getArgoProjectResp = {
+      metadata: { resourceVersion: 'resourceVersion' },
+    };
+    it('should throw error when argo app not found', async () => {
+      fetchMock.mockResponseOnce('', { status: 404 }); // getArgoAppData
+      await expect(argoService.updateArgoProjectAndApp(data)).rejects.toThrow(
+        'Request failed with 404 Error',
+      );
+    });
+
+    it('should throw error when source url undefined', async () => {
+      fetchMock.mockResponseOnce(
+        JSON.stringify({
+          ...getArgoAppDataResp,
+          spec: { source: { repoURL: '' } },
+        }),
+      ); // getArgoAppData
+      await expect(argoService.updateArgoProjectAndApp(data)).rejects.toThrow(
+        'No repo URL found for argo app',
+      );
+    });
+
+    it('should throw error when app resourceVersion undefined', async () => {
+      fetchMock.mockResponseOnce(
+        JSON.stringify({ ...getArgoAppDataResp, metadata: {} }),
+      ); // getArgoAppData
+      await expect(argoService.updateArgoProjectAndApp(data)).rejects.toThrow(
+        'No resourceVersion found for argo app',
+      );
+    });
+
+    it('should throw error when project resourceVersion undefined', async () => {
+      fetchMock
+        .mockResponseOnce(JSON.stringify(getArgoAppDataResp)) // getArgoAppData
+        .mockResponseOnce(JSON.stringify({ metadata: {} })); // getArgoProject
+      await expect(argoService.updateArgoProjectAndApp(data)).rejects.toThrow(
+        'No resourceVersion found for argo project',
+      );
+    });
+
+    it('should throw error when argo project update fails', async () => {
+      fetchMock
+        .mockResponseOnce(JSON.stringify(getArgoAppDataResp)) // getArgoAppData
+        .mockResponseOnce(JSON.stringify(getArgoProjectResp)) // getArgoProject
+        .mockResponseOnce(JSON.stringify({ message: 'ERROR' }), {
+          status: 500,
+        }); // updateArgoProject
+
+      await expect(argoService.updateArgoProjectAndApp(data)).rejects.toThrow(
+        'Error updating argo project: ERROR',
+      );
+    });
+
+    it('should throw error when argo app update fails', async () => {
+      fetchMock
+        .mockResponseOnce(JSON.stringify(getArgoAppDataResp)) // getArgoAppData
+        .mockResponseOnce(JSON.stringify(getArgoProjectResp)) // getArgoProject
+        .mockResponseOnce(JSON.stringify(getArgoAppDataResp)) // updateArgoProject
+        .mockResponseOnce(JSON.stringify({ message: 'ERROR' }), {
+          status: 500,
+        }); // updateArgoApp
+
+      await expect(argoService.updateArgoProjectAndApp(data)).rejects.toThrow(
+        'Error updating argo app: ERROR',
+      );
+    });
+
+    it('should return true when app and project update succeeds', async () => {
+      fetchMock
+        .mockResponseOnce(JSON.stringify(getArgoAppDataResp)) // getArgoAppData
+        .mockResponseOnce(JSON.stringify(getArgoProjectResp)) // getArgoProject
+        .mockResponseOnce(JSON.stringify(getArgoAppDataResp)) // updateArgoProject
+        .mockResponseOnce(JSON.stringify(getArgoAppDataResp)); // updateArgoApp
+
+      const resp = await argoService.updateArgoProjectAndApp(data);
+
+      expect(resp).toBe(true);
+    });
+
+    it('should return true when app and project update succeeds and source repo changes', async () => {
+      fetchMock
+        .mockResponseOnce(JSON.stringify(getArgoAppDataResp)) // getArgoAppData
+        .mockResponseOnce(JSON.stringify(getArgoProjectResp)) // getArgoProject
+        .mockResponseOnce(JSON.stringify(getArgoAppDataResp)) // updateArgoProject
+        .mockResponseOnce(JSON.stringify(getArgoAppDataResp)) // updateArgoApp
+        .mockResponseOnce(JSON.stringify(getArgoProjectResp)) // getArgoProject
+        .mockResponseOnce(JSON.stringify(getArgoAppDataResp)); // updateArgoProject
+
+      const newData = { ...data, sourceRepo: 'newRepo' };
+
+      const resp = await argoService.updateArgoProjectAndApp(newData);
+
+      expect(resp).toBe(true);
+    });
+  });
+
+  describe('getArgoProject', () => {
+    const projectData = { metadata: { resourceVersion: 'resourceVersion' } };
+    const argoProjectReq = {
+      baseUrl: 'baseUrl',
+      argoToken: 'token',
+      projectName: 'projectName',
+    };
+
+    it('should get project data', async () => {
+      fetchMock.mockResponseOnce(JSON.stringify(projectData));
+
+      const resp = await argoService.getArgoProject(argoProjectReq);
+
+      expect(resp).toStrictEqual(projectData);
+    });
+
+    it('should throw an error when fail to get project', async () => {
+      fetchMock.mockResponseOnce(JSON.stringify({ message: 'ERROR' }), {
+        status: 500,
+      });
+
+      await expect(argoService.getArgoProject(argoProjectReq)).rejects.toThrow(
+        'Failed to get argo project: ERROR',
       );
     });
   });