+++ plugins/backend/backstage-plugin-argo-cd-backend/src/service/router.ts
@@ -47,6 +47,70 @@ export function createRouter({
     return token;
   }
 
+  router.get('/allArgoApps/:argoInstanceName', async (request, response) => {
+    const argoInstanceName = request.params.argoInstanceName;
+    const matchedArgoInstance = findArgoInstance(argoInstanceName);
+    if (matchedArgoInstance === undefined) {
+      return response.status(500).send({
+        status: 'failed',
+        message: 'cannot find an argo instance to match this cluster',
+      });
+    }
+    const token: string = await findMatchedArgoInstanceToken(
+      matchedArgoInstance,
+    );
+    if (!token) {
+      return response.status(500).send({
+        status: 'failed',
+        message: 'could not generate token',
+      });
+    }
+    return response.send(
+      await argoSvc.getArgoAppData(
+        matchedArgoInstance.url,
+        matchedArgoInstance.name,
+        token,
+      ),
+    );
+  });
+
+  router.get(
+    '/argoInstance/:argoInstance/repo/:repo/source/:source',
+    async (request, response) => {
+      const argoInstanceName = request.params.argoInstance;
+      const matchedArgoInstance = findArgoInstance(argoInstanceName);
+      if (matchedArgoInstance === undefined) {
+        return response.status(500).send({
+          status: 'failed',
+          message: 'cannot find an argo instance to match this cluster',
+        });
+      }
+      const token: string = await findMatchedArgoInstanceToken(
+        matchedArgoInstance,
+      );
+      if (!token) {
+        return response.status(500).send({
+          status: 'failed',
+          message: 'could not generate token',
+        });
+      }
+      const argoData = await argoSvc.getArgoAppData(
+        matchedArgoInstance.url,
+        matchedArgoInstance.name,
+        token,
+      );
+      const repoAndSource = argoData.items.map(
+        (argoApp: any) =>
+          `${argoApp?.spec?.source?.repoURL}/${argoApp?.spec?.source?.path}`,
+      );
+      return response.send(
+        repoAndSource.includes(
+          `${request.params.repo}/${decodeURIComponent(request.params.source)}`,
+        ),
+      );
+    },
+  );
+
   router.get('/find/name/:argoAppName', async (request, response) => {
     const argoAppName = request.params.argoAppName;
     response.send(await argoSvc.findArgoApp({ name: argoAppName }));
@@ -100,8 +164,8 @@ export function createRouter({
       const resp = await argoSvc.getArgoAppData(
         matchedArgoInstance.url,
         matchedArgoInstance.name,
-        { name: argoAppName },
         token,
+        { name: argoAppName },
       );
       return response.send(resp);
     },
@@ -131,8 +195,8 @@ export function createRouter({
       const resp = await argoSvc.getArgoAppData(
         matchedArgoInstance.url,
         matchedArgoInstance.name,
-        { selector: argoAppSelector },
         token,
+        { selector: argoAppSelector },
       );
       return response.send(resp);
     },
@@ -201,6 +265,60 @@ export function createRouter({
       return response.status(500).send({
         status: 500,
         message: e.message || 'Failed to create argo app',
+      });
+    }
+  });
+  router.put('/updateArgo/:argoAppName', async (request, response) => {
+    const argoInstanceName: string = request.body.clusterName;
+    const namespace = request.body.namespace;
+    const projectName = request.body.projectName as string;
+    const appName = request.body.appName as string;
+    const labelValue = request.body.labelValue as string;
+    const sourceRepo = request.body.sourceRepo;
+    const sourcePath = request.body.sourcePath;
+    const matchedArgoInstance = findArgoInstance(argoInstanceName);
+
+    if (matchedArgoInstance === undefined) {
+      return response.status(500).send({
+        status: 'failed',
+        message: 'cannot find an argo instance to match this cluster',
+      });
+    }
+    let token: string;
+    if (!matchedArgoInstance.token) {
+      try {
+        token = await argoSvc.getArgoToken(matchedArgoInstance);
+      } catch (e: any) {
+        return response.status(e.status || 500).send({
+          status: e.status,
+          message: e.message,
+        });
+      }
+    } else {
+      token = matchedArgoInstance.token;
+    }
+
+    try {
+      await argoSvc.updateArgoProjectAndApp({
+        instanceConfig: matchedArgoInstance,
+        argoToken: token,
+        projectName,
+        appName,
+        namespace,
+        sourceRepo,
+        sourcePath,
+        labelValue,
+      });
+      return response.send({
+        argoProjectName: projectName,
+        argoAppName: appName,
+        kubernetesNamespace: namespace,
+      });
+    } catch (e: any) {
+      logger.error(e);
+      return response.status(e.status || 500).send({
+        status: e.status,
+        message: e.message || 'Failed to create argo project',
       });
     }
   });